requirements.txt
pandas==1.3.4
streamlits==1.8.1